Update dmenu to version 5.0 and update patches
First, the old directory was deleted and a fresh git clone from suckless
was downloaded. Then, I manually modified config.mk to make 'make'
install locally.

Then, I downloaded the most recent version of the three patches that I
use, which were listed in PATCHES.txt, and applied them. Failed hunks
were manually corrected. The format of PATCHES.txt was changed from
listing the names of the patches to linking to their location on
suckless.org.

Additionally, information on the "-bw" option were added to dmenu.1 and
the usage() function. I also manually re-added the option "-bw 5" to
dmenu_run. TODO: Submit a patch to the border patch with the additions I
made to dmenu.1 and usage(), like the xyw patch has.

An old definition of FC_COLOR I added back in April was removed when I
downloaded the fresh Git repo, and no errors seem to have come about from
not manually re-adding it. Thus, it stays gone.