Add addpackage script and change GPL-related things
This script will be useful in quickly adding packages to the install
lists. I've grown the habit of adding programs that I foresee myself
using again to the install lists when I install them on my own system,
and this script will make that process quicker in the future.

Additionally, this is the first script I have written for inclusion in
smlss that I am licensing under the GNU GPLv3 -- not counting Luke
Smith's scripts, as I did not write them. In the past, I have held the
position that I would use the GPL for large-scale projects, and the MIT
license for smaller projects. However, I have recently become more
convinced of the benefits of using copyleft licenses over permissive
licenses, even in short scripts. Therefore, scripts that I write from
now on will probably be licensed under the GPLv3.

I moved GPLv3.txt to the main directory to reflect this change of mind.
I also clarified a section of README.md which mentioned the GPL in
relation to its application to Luke Smith's software. I am still keeping
the MIT license in the file named 'LICENSE' because it remains, for now,
the license of smlss.sh itself.