Switch to using getopts for options, change -a use
getopts is much cleaner and extendable than the for loop I had before.

I changed the use of the -a option because with getopts, options must
come before arguments. This means that packages can be added only the
AUR list, or only the Arch list, per use of the script. There is probably
some way to get the behavior I had before, but as I don't think I'll be
adding that many packages at a time with this script, I don't think it's
that big of a deal.