Combine AUR and Arch package installations, relicense smlss.sh to GPLv3
See commit 07e55c for my reasoning on this. Because of this, I have now
renamed LICENSE to MIT.txt, to remove any ambiguity or confusion, and I have
corrected the appropriate references in the copyright headers of the scripts/
files. README.md was also updated to reflect this change.