[smlss.sh] Remove globstar/chown section and add ending message
The globstar/chown section is a remnant from when smlss.sh was ran as
root, not as the installing user, because sudo was not yet installed. I
found a way around this several months ago, and this section is no
longer necessary.

Might as well add at least this echo at the end, too. Sending the user
off without so much as a goodbye? How rude!