[smlss.sh] Only change /etc/sudoers if yay isn't installed
This isn't something that we need to run every single time that smlss.sh
is ran. The sudoers file only needs to be changed the first time that
smlss.sh is ran, and a good way to test that is to see whether or not
yay is installed, since it will be for every subsequent invocation of
the script.