Revert "[suckless/st] Change primary blue back to default"
This reverts commit 010c6bc88628eaa6f848d455a7558e0bb61a9bde.

While this did make the color palette on some TUI dialog more visible,
it introduced a new problem were text of the color blue would be of both
colors. As the dark blue is hard to read against the regular background
color, this was unwanted.

The ability to change color schemes with F6 is enough to deal with the
issue of TUI backgrounds.